+++ RobotPosition.h
@@ -11,11 +11,11 @@ public:
 	virtual void updatePolarCoords();
 	virtual void updateFieldCoords(/*GatePosition yellowGate, GatePosition blueGate*/);
 	virtual void updateCoordinates(/*GatePosition yellowGate, GatePosition blueGate*/);
-	int getAngle();
+	double getAngle();
 private:
 	GatePosition & yellowGate, & blueGate; // use references that point somewhere
 	void initPolarCoordinates();
 	std::pair<cv::Point, cv::Point> intersectionOfTwoCircles(cv::Point circle1center, double circle1Rad, cv::Point circle2center, double circle2Rad);
-	bool isRobotAboveCenterLine(int yellowGoalAngle, int blueGoalAngle);
-	int getRobotDirection(int yellowGoalDist, int yellowGoalAngle, int blueGoalDist, int blueGoalAngle);
+	bool isRobotAboveCenterLine(double yellowGoalAngle, double blueGoalAngle);
+	double getRobotDirection(double yellowGoalDist, double yellowGoalAngle, double blueGoalDist, double blueGoalAngle);
 };
